Moved IntroActivity.Model to IntroModel
