+++ app/src/main/kotlin/at/bitfire/davdroid/ui/intro/IntroActivity.kt
@@ -16,19 +16,15 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.pager.rememberPagerState
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.lifecycle.ViewModel
-import at.bitfire.davdroid.log.Logger
 import at.bitfire.davdroid.ui.AppTheme
 import dagger.hilt.android.AndroidEntryPoint
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
 import kotlinx.coroutines.launch
 
 @AndroidEntryPoint
 @OptIn(ExperimentalFoundationApi::class)
 class IntroActivity : AppCompatActivity() {
 
-    val model by viewModels<Model>()
+    val model by viewModels<IntroModel>()
 
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -75,51 +71,6 @@ class IntroActivity : AppCompatActivity() {
         override fun parseResult(resultCode: Int, intent: Intent?): Boolean {
             return resultCode == Activity.RESULT_CANCELED
         }
-    }
-
-
-    @HiltViewModel
-    class Model @Inject constructor(
-        introPageFactory: IntroPageFactory
-    ): ViewModel() {
-
-        private val introPages = introPageFactory.introPages
-
-        private var _pages: List<IntroPage>? = null
-        val pages: List<IntroPage>
-            @Synchronized
-            get() {
-                _pages?.let { return it }
-
-                val newPages = calculatePages()
-                _pages = newPages
-
-                return newPages
-            }
-
-        private fun calculatePages(): List<IntroPage> {
-            for (page in introPages)
-                Logger.log.fine("Found intro page ${page::class.java} with order ${page.getShowPolicy()}")
-
-            val activePages: Map<IntroPage, IntroPage.ShowPolicy> = introPages
-                .associateWith { page ->
-                    page.getShowPolicy().also { policy ->
-                        Logger.log.fine("IntroActivity: found intro page ${page::class.java} with $policy")
-                    }
-                }
-                .filterValues { it != IntroPage.ShowPolicy.DONT_SHOW }
-
-            val anyShowAlways = activePages.values.any { it == IntroPage.ShowPolicy.SHOW_ALWAYS }
-            return if (anyShowAlways) {
-                val pages = mutableListOf<IntroPage>()
-                activePages.filterValues { it != IntroPage.ShowPolicy.DONT_SHOW }.forEach { page, _ ->
-                    pages += page
-                }
-                pages
-            } else
-                emptyList()
-        }
-
     }
 
 }
